flash_wls_sensor_app/requirements.txt
customtkinter==5.2.2
pillow>=12.0.0
pymcuprog>=3.19
pyserial==3.5
